corrected got pep8 problem: /Users/Piotr/PycharmProjects/lunch-app/src/lunch_app/views.py:500:38: E712 comparison to False should be 'if cond is False:' or 'if not cond:'
to do is test checking for popularity of the meal
