+++ src/lunch_app/forms.py
@@ -142,21 +142,16 @@ class AddFood(Form):
         Form.__init__(self, *args, **kwargs)
 
 
-class DidUserPayForm(Form):
-    """
-    Did user Pay bool field
-    """
-    did_user_pay = SelectField(
-        'did_user_pay',
-        choices=[('1', 'Tak'), ('0', 'Nie')],
-        default='0',
-    )
-
-
 class MailTextForm(Form):
     """
     Did user Pay bool field
     """
+    daily_reminder_subject = TextAreaField(
+        "daily_reminder_subject",
+        validators=[validators.DataRequired(
+            "Please enter daily reminder text."
+        )]
+    )
     daily_reminder = TextAreaField(
         "daily_reminder",
         validators=[validators.DataRequired(
@@ -185,7 +180,7 @@ class MailTextForm(Form):
 
 class UserDailyReminderForm(Form):
     """
-    Did user Pay field
+    User daily reminder form.
     """
     i_want_daily_reminder = BooleanField(
         'i_want_daily_reminder',
@@ -208,4 +203,4 @@ class FinanceSearchForm(Form):
         ('0', 'All'),
         ('1', 'Paid'),
         ('2', 'Unpaid'),
-        ])
+    ])
