+++ src/lunch_app/templates/base.html
@@ -32,6 +32,7 @@
             <ul class="title-area">
                 <h1>STX Lunch</h1>
 
+
                 {% if current_user and current_user.is_authenticated() %}
                     <nav class="top-bar" data-topbar role="navigation">
