+++ src/lunch_app/templates/finance_base.html
@@ -1,16 +1,16 @@
 <dl class="sub-nav">
-    <dt>Finance:</dt>
+    <dt>Finance</dt>
 
     <dd {% if page_id == 'finance_main' %}
         class="active" {% endif %}><a href="{{ url_for('finance', year=get_current_year(), month=get_current_month(), did_pay=0) }}">Summary</a></dd>
     <dd {% if page_id == 'finance_unpaid' %}
         class="active" {% endif %}><a href="{{ url_for('finance', year=get_current_year(), month=get_current_month(), did_pay=1) }}">Paid</a></dd>
     <dd {% if page_id == 'finance_unpaid' %}
-        class="active" {% endif %}><a href="{{ url_for('finance', year=get_current_year(), month=get_current_month(), did_pay=2) }}">Not Paid</a></dd>
+        class="active" {% endif %}><a href="{{ url_for('finance', year=get_current_year(), month=get_current_month(), did_pay=2) }}">Not paid</a></dd>
     <dd {% if page_id == 'finance_mail_text' %}
-        class="active" {% endif %}><a href="{{ url_for('finance_mail_text') }}">Mail Text</a></dd>
+        class="active" {% endif %}><a href="{{ url_for('finance_mail_text') }}">Mail text</a></dd>
     <dd {% if page_id == 'finance_mail_all' %}
-        class="active" {% endif %}><a href="{{ url_for('finance_mail_all') }}">Mail All</a></dd>
+        class="active" {% endif %}><a href="{{ url_for('finance_mail_all') }}">Mail all</a></dd>
     <dd {% if page_id == 'finance_search' %}
         class="active" {% endif %}><a href="{{ url_for('finance_search_view') }}">Search</a></dd>
