+++ src/lunch_app/templates/finance_mail_all.html
@@ -16,7 +16,7 @@
     </div>
     <div class="large-6 columns">
         <form method="POST">
-            Click to send mail to every one with their monthly summary.
+            Click to send mail to everyone with their monthly summary.
             <input type="submit" value="submit" class="button right">
 
         </form>
