+++ src/lunch_app/templates/finance_mail_text.html
@@ -8,6 +8,8 @@
 
     <div class="large-6 columns">
 <form method="POST">
+            Daily reminder subject
+            {{ form.daily_reminder_subject }}
             Daily reminder text
             {{ form.daily_reminder }}
             Month summary text
