+++ src/lunch_app/templates/finance_search.html
@@ -14,7 +14,7 @@
             {{ form.year }}
             <b>Month</b>
             {{ form.month }}
-            <b>All/Paid/Unpaid Users</b>
+            <b>All/Paid/Unpaid users</b>
             {{ form.did_pay }}
 
             <input type="submit" value="submit" class="button right">
