+++ src/lunch_app/templates/overview.html
@@ -12,11 +12,12 @@
 
                 <form method="POST">
 
-                    I want to recive email evry day at 10:45 wtih reminder if I
-                    have <b>not</b> ordered that day.
+                    I desire receiving email every day at 10:45 with reminder
+                    that I did <b>not</b> commit order on that marvelous day.
 
 
                     {{ form.i_want_daily_reminder }}
+                    <br>
 
                     <input type="submit" value="submit" class="button right">
             </div>
